requirements.txt
rpaframework
pandas==2.0.3
python_dateutil==2.8.2
Requests==2.31.0
selenium==4.10.0